+++ .pre-commit-config.yaml
@@ -1,7 +1,7 @@
 # Configuration file for `pre-commit` tool.
 # Source: https://pre-commit.com/#pre-commit-configyaml---top-level
 exclude: "docs|node_modules|migrations|.git|.tox|README.md|sample_setups/(external-setups|jsons)|tests/deprecated"
-default_stages: [commit]
+default_stages: [pre-commit]
 fail_fast: true
 
 repos:
